// tools/conformance/src/modules/nestjs-command-module-generator/nestjs-command-module-generator.types.ts
// 🏷️ Types

import type { Tree } from "@nx/devkit";

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsCommandModuleGeneratorArguments {
  options: NestjsCommandModuleGeneratorOptions;
  tree: Tree;
}

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsCommandModuleGeneratorOptions {
  name: string;
  project?: string;
}